Finish initial version of hardware abstraction
