+++ backend/package.json
@@ -50,6 +50,7 @@
     "omggif": "^1.0.10",
     "pngjs": "^7.0.0",
     "reflect-metadata": "^0.2.2",
+    "rxjs": "^7.8.2",
     "serialport": "^12.0.0",
     "sharp": "^0.33.5",
     "socket.io": "^4.8.0",
