+++ backend/src/config/ConfigLoader.ts
@@ -2,6 +2,7 @@ import { Hardware } from "../hardware/Hardware";
 import { Configuration } from "./objects/Configurations";
 import { DatabaseConfiguration } from "./objects/DatabaseConfiguration";
 import { FlaschenTaschenConfiguration } from "./objects/FlaschenTaschenConfiguration";
+import { HUDConfiguration } from "./objects/HUDConfiguration";
 import { LedMatrixConfiguration } from "./objects/LedMatrixConfiguration";
 import { MiscConfiguration } from "./objects/MiscConfiguration";
 import { RemoteWorkerConfiguration } from "./objects/RemoteWorkerConfiguration";
@@ -147,17 +148,12 @@ export function loadConfiguration(): Configuration {
   }
   //#endregion
 
-  //#region Arduino
+  //#region Serial
   const serialPort = process.env["SERIAL_PORT"];
   const serialBaudRate = parseInt(String(process.env["SERIAL_BAUD_RATE"]));
-  const oledTextLines = parseInt(String(process.env["OLED_TEXT_LINES"]));
 
   if (isNaN(serialBaudRate) || serialBaudRate <= 0) {
     throw new Error("Missing or invalid: SERIAL_BAUD_RATE");
-  }
-
-  if (isNaN(oledTextLines) || oledTextLines <= 0) {
-    throw new Error("Missing or invalid: OLED_TEXT_LINES");
   }
 
   if (serialPort == null) {
@@ -167,7 +163,18 @@ export function loadConfiguration(): Configuration {
   const serialConfig: SerialConfiguration = {
     port: serialPort,
     baudRate: serialBaudRate,
-    oledTextLines: oledTextLines,
+  }
+  //#endregion
+
+  //#region HUD
+  const oledTextLines = parseInt(String(process.env["OLED_TEXT_LINES"]));
+
+  if (isNaN(oledTextLines) || oledTextLines <= 0) {
+    throw new Error("Missing or invalid: OLED_TEXT_LINES");
+  }
+
+  const hudConfig: HUDConfiguration = {
+    lines: oledTextLines,
   }
   //#endregion
 
@@ -214,6 +221,7 @@ export function loadConfiguration(): Configuration {
     logDirectory: logDirectory,
     misc: misc,
     hardware: hardware,
+    hud: hudConfig,
   }
 }
