+++ backend/src/config/objects/HUDConfiguration.ts
@@ -1,0 +1,3 @@
+export interface HUDConfiguration {
+  lines: number;
+}
