+++ backend/src/config/objects/SerialConfiguration.ts
@@ -1,5 +1,4 @@
 export interface SerialConfiguration {
   port: string;
   baudRate: number;
-  oledTextLines: number;
 }
